Refactor lexer tests by splitting into thematic files: LexerCommentTest, LexerOperatorTest, LexerExpressionTest, and LexerKeywordTest for improved readability and maintainability.
